refactor: linkerhook 重写为 C 并合并驱动加载到 egl_bridge，与 Amethyst-Android 布局一致
