--- FCLauncher/src/main/jni/driver_helper/driver_helper.h
@@ -1,6 +1,0 @@
-#ifndef FCL_DRIVER_HELPER_H
-#define FCL_DRIVER_HELPER_H
-
-void load_vulkan();
-
-#endif //FCL_DRIVER_HELPER_H
