+++ FCLauncher/src/main/jni/egl_bridge.c
@@ -13,6 +13,7 @@
 
 #include <EGL/egl.h>
 #include "GL/osmesa.h"
+#include "GL/gl.h"
 #include "ctxbridges/osmesa_loader.h"
 #include "ctxbridges/egl_loader.h"
 #include "virgl/virgl.h"
@@ -25,11 +26,13 @@
 #include <android/native_window_jni.h>
 #include <android/rect.h>
 #include <string.h>
+#include <inttypes.h>
 #include "environ/environ.h"
 #include <android/dlext.h>
 #include "ctxbridges/bridge_tbl.h"
 #include "ctxbridges/osm_bridge.h"
-#include "driver_helper/driver_helper.h"
+#include "driver_helper/nsbypass.h"
+#include "fcl/include/fcl_internal.h"
 #include <stdatomic.h>
 
 // 由 input_bridge_v3.c 提供，上报 monitor size 到 Java 侧 GLFW
@@ -101,6 +104,131 @@ EXTERNAL_API void *pojavGetCurrentContext() {
         return virglGetCurrentContext();
     }
     return br_get_current();
+}
+
+#ifdef ADRENO_POSSIBLE
+
+bool checkAdrenoGraphics() {
+    EGLDisplay eglDisplay = eglGetDisplay(EGL_DEFAULT_DISPLAY);
+    if (eglDisplay == EGL_NO_DISPLAY || eglInitialize(eglDisplay, NULL, NULL) != EGL_TRUE)
+        return false;
+
+    EGLint egl_attributes[] = {
+        EGL_BLUE_SIZE, 8, EGL_GREEN_SIZE, 8, EGL_RED_SIZE, 8,
+        EGL_ALPHA_SIZE, 8, EGL_DEPTH_SIZE, 24, EGL_SURFACE_TYPE, EGL_PBUFFER_BIT,
+        EGL_RENDERABLE_TYPE, EGL_OPENGL_ES2_BIT, EGL_NONE
+    };
+
+    EGLint num_configs = 0;
+    if (eglChooseConfig(eglDisplay, egl_attributes, NULL, 0, &num_configs) != EGL_TRUE || num_configs == 0) {
+        eglTerminate(eglDisplay);
+        return false;
+    }
+
+    EGLConfig eglConfig;
+    eglChooseConfig(eglDisplay, egl_attributes, &eglConfig, 1, &num_configs);
+
+    const EGLint egl_context_attributes[] = { EGL_CONTEXT_CLIENT_VERSION, 3, EGL_NONE };
+    EGLContext context = eglCreateContext(eglDisplay, eglConfig, EGL_NO_CONTEXT, egl_context_attributes);
+    if (context == EGL_NO_CONTEXT) {
+        eglTerminate(eglDisplay);
+        return false;
+    }
+
+    if (eglMakeCurrent(eglDisplay, EGL_NO_SURFACE, EGL_NO_SURFACE, context) != EGL_TRUE) {
+        eglDestroyContext(eglDisplay, context);
+        eglTerminate(eglDisplay);
+        return false;
+    }
+
+    const char* vendor = (const char*)glGetString(GL_VENDOR);
+    const char* renderer = (const char*)glGetString(GL_RENDERER);
+
+    bool is_adreno = (vendor && renderer && strcmp(vendor, "Qualcomm") == 0 && strstr(renderer, "Adreno") != NULL);
+
+    eglMakeCurrent(eglDisplay, EGL_NO_SURFACE, EGL_NO_SURFACE, EGL_NO_CONTEXT);
+    eglDestroyContext(eglDisplay, context);
+    eglTerminate(eglDisplay);
+
+    return is_adreno;
+}
+
+void* loadTurnipVulkan() {
+    if (!checkAdrenoGraphics())
+        return NULL;
+
+    const char* native_dir = getenv("DRIVER_PATH");
+    const char* cache_dir = getenv("TMPDIR");
+
+    if (!native_dir)
+        return NULL;
+
+    if (!linker_ns_load(native_dir))
+        return NULL;
+
+    void* linkerhook = linker_ns_dlopen("liblinkerhook.so", RTLD_LOCAL | RTLD_NOW);
+    if (!linkerhook)
+        return NULL;
+
+    void* turnip_driver_handle = linker_ns_dlopen("libvulkan_freedreno.so", RTLD_LOCAL | RTLD_NOW);
+    if (!turnip_driver_handle) {
+        dlclose(linkerhook);
+        return NULL;
+    }
+
+    void* dl_android = linker_ns_dlopen("libdl_android.so", RTLD_LOCAL | RTLD_LAZY);
+    if (!dl_android) {
+        dlclose(linkerhook);
+        dlclose(turnip_driver_handle);
+        return NULL;
+    }
+
+    void* android_get_exported_namespace = dlsym(dl_android, "android_get_exported_namespace");
+    void (*linkerhookPassHandles)(void*, void*, void*) = dlsym(linkerhook, "app__pojav_linkerhook_pass_handles");
+
+    if (!linkerhookPassHandles || !android_get_exported_namespace) {
+        dlclose(dl_android);
+        dlclose(linkerhook);
+        dlclose(turnip_driver_handle);
+        return NULL;
+    }
+
+    linkerhookPassHandles(turnip_driver_handle, android_dlopen_ext, android_get_exported_namespace);
+
+    void* libvulkan = linker_ns_dlopen_unique(cache_dir, "libvulkan.so", RTLD_LOCAL | RTLD_NOW);
+    if (!libvulkan) {
+        dlclose(dl_android);
+        dlclose(linkerhook);
+        dlclose(turnip_driver_handle);
+        return NULL;
+    }
+
+    return libvulkan;
+}
+
+#endif
+
+static void set_vulkan_ptr(void* ptr) {
+    char envval[64];
+    sprintf(envval, "%"PRIxPTR, (uintptr_t)ptr);
+    setenv("VULKAN_PTR", envval, 1);
+}
+
+void load_vulkan() {
+    if(getenv("VULKAN_DRIVER_SYSTEM") == NULL && android_get_device_api_level() >= 28) {
+#ifdef ADRENO_POSSIBLE
+        void* result = loadTurnipVulkan();
+        if(result != NULL) {
+            FCL_LOG("AdrenoSupp: Loaded Turnip, loader address: %p", result);
+            set_vulkan_ptr(result);
+            return;
+        }
+#endif
+    }
+    FCL_LOG("OSMDroid: loading vulkan regularly...");
+    void* vulkan_ptr = dlopen("libvulkan.so", RTLD_LAZY | RTLD_LOCAL);
+    FCL_LOG("OSMDroid: loaded vulkan, ptr=%p", vulkan_ptr);
+    set_vulkan_ptr(vulkan_ptr);
 }
 
 int pojavInitOpenGL() {
